+++ examples/docs_snippets/docs_snippets/overview/configuration/configured_solid_example.py
@@ -9,10 +9,14 @@ def example_solid(context):
         context.log.info(context.solid_config["word"])
 
 
-# returns a solid named 'example_solid'
-new_solid = configured(example_solid)({"iterations": 6, "word": "wheaties"})
-
-# returns a solid named 'configured_example_solid'
-another_new_solid = configured(example_solid, name="configured_example")(
+# This example is fully configured. With this syntax, a name must be explicitly provided.
+configured_example = configured(example_solid, name="configured_example")(
     {"iterations": 6, "word": "wheaties"}
 )
+
+# This example is partially configured: `iterations` is passed through
+# The decorator yields a solid named 'another_configured_example' (from the decorated function)
+# with `int` as the `config_schema`.
+@configured(example_solid, int)
+def another_configured_example(config):
+    return {"iterations": config, "word": "wheaties"}