+++ examples/docs_snippets/docs_snippets_tests/overview_tests/configuration_tests/test_configured_solid_example.py
@@ -1,20 +1,28 @@
 from docs_snippets.overview.configuration.configured_solid_example import (
-    another_new_solid,
-    new_solid,
+    another_configured_example,
+    configured_example,
 )
 
 from dagster.utils.test import execute_solid
 
 
 def test_new_solid(capsys):
-    execute_solid(new_solid)
+    assert configured_example.name == "configured_example"
+    execute_solid(configured_example)
 
     captured = capsys.readouterr()
     assert captured.err.count("wheaties") == 6
 
 
 def test_another_new_solid(capsys):
-    execute_solid(another_new_solid)
+    assert another_configured_example.name == "another_configured_example"
+    execute_solid(
+        another_configured_example,
+        None,
+        None,
+        None,
+        {"solids": {"another_configured_example": {"config": 6}}},
+    )
 
     captured = capsys.readouterr()
     assert captured.err.count("wheaties") == 6